pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-compression-experiments"
version = "0.1.0"
description = "Context compression optimization for Agentic RAG using DSPy GEPA"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Laurian Gridinoc", email = "laurian@gridinoc.com"},
]
keywords = [
    "context-compression",
    "rag", 
    "dspy",
    "gepa",
    "llm",
    "optimization"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core ML/NLP libraries
    "dspy-ai>=2.4.0",
    "textgrad>=0.1.4",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "loguru>=0.7.0",
    
    # Analysis and visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    
    # Experiment tracking
    "wandb>=0.16.0",
    "weave>=0.50.0",
    
    # Jupyter for experimentation
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.0.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/laurian/context-compression-experiments-2508"
Repository = "https://github.com/laurian/context-compression-experiments-2508"
Issues = "https://github.com/laurian/context-compression-experiments-2508/issues"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = [
    "tests",
]